refactor: moved GetTask to task file
